requirements.txt
aiohttp==3.9.5
aiosignal==1.4.0
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
python-dotenv==1.2.2
requests==2.32.3
typing_extensions==4.15.0
urllib3==2.2.1
yarl==1.23.0
